+++ examples/relay/inbound_dial.rb
@@ -15,7 +15,7 @@ class MyConsumer < Signalwire::Relay::Consumer
 
   def on_incoming_call(call)
     call.answer
-    call.play [{ "type": 'tts', "params": { "text": 'connecting you to the clock service', "language": 'en-US', "gender": 'male' } }]
+    call.play_tts 'connecting you to the clock service'
     call.connect [[{ type: 'phone', params: { to_number: '+12027621401', from_number: ENV['FROM_NUMBER'], timeout: 30 } }]]
     sleep 20
     call.hangup