+++ examples/relay/outbound_collect.rb
@@ -8,19 +8,12 @@ $LOAD_PATH.unshift(File.dirname(__FILE__) + '/../lib')
 
 class OutboundConsumer < Signalwire::Relay::Consumer
   def ready
-    call = client.calling.new_call(from: ENV['FROM_NUMBER'], to: ENV['TO_NUMBER'])
-    call.dial
+    call = client.calling.new_call(from: ENV['FROM_NUMBER'], to: ENV['TO_NUMBER']).dial
     collect_params = { "initial_timeout": 10.0, "digits": { "max": 1, "digit_timeout": 5.0 } }
-    play_params = [{ "type": 'tts', "params": { "text": 'how many hamburgers would you like to order?', "language": 'en-US', "gender": 'male' } }]
-    puts "prompting"
-    result = call.prompt( collect_params, play_params)
+    result = call.prompt_tts( collect_params, 'how many hamburgers would you like to order?')
 
-    puts result.result
-    call.play [{ "type": 'tts', "params": { "text": "You ordered #{result.result} hamburgers. Thank you!", "language": 'en-US', "gender": 'male' } }]
+    call.play_tts "You ordered #{result.result} hamburgers. Thank you!"
     call.hangup
-  rescue Exception => e
-    puts e.inspect
-    puts e.backtrace
   end
 end