+++ lib/signalwire/rest/client.rb
@@ -3,11 +3,11 @@
 module Signalwire::REST
   class Client < Twilio::REST::Client
     def initialize(username = nil, password = nil, account_sid = nil, region = nil, http_client = Twilio::HTTP::Client.new, **args)
-      signalwire_space_url = args.delete(:signalwire_space_url)
+      host = args.delete(:SIGNALWIRE_HOST)
 
-      unless signalwire_space_url.nil?
+      unless host.nil?
         Signalwire::Sdk.configure do |config|
-          config.hostname = signalwire_space_url
+          config.hostname = host
         end
       end