+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+## 2.0.0-rc.1
+
+- Relay client
+
 ## 1.4.0
 
 - Add ability to specify domain via parameter
-- Accept both `SIGNALWIRE_SPACE_URL` and `SIGNALWIRE_API_HOSTNAME` variables for configuration
+- Accept both `SIGNALWIRE_HOST` and `SIGNALWIRE_API_HOSTNAME` variables for configuration
 
 ## 1.3.0